+++ src/biz/bokhorst/xprivacy/ActivityMain.java
@@ -1110,17 +1110,16 @@ public class ActivityMain extends ActivityBase implements OnItemSelectedListener
 				convertView = mInflater.inflate(R.layout.templateentry, null);
 				holder = new ViewHolder(convertView, position);
 				convertView.setTag(holder);
-			} else {
+			} else
 				holder = (ViewHolder) convertView.getTag();
-			}
 
 			// Get info
 			final String templateName = PrivacyManager.cSettingTemplate + "." + listRestrictionName.get(position);
-			String value = PrivacyManager.getSetting(0, templateName, "false+ask", false);
+			String value = PrivacyManager.getSetting(0, templateName, Boolean.toString(!ondemand) + "+ask", false);
 			holder.restricted = value.contains("true");
-			holder.asked = !ondemand || value.contains("asked");
-			Bitmap check = holder.asked ? (holder.restricted ? getFullCheckBox() : getOffCheckBox())
-					: getOnDemandCheckBox();
+			holder.asked = (!ondemand || value.contains("asked"));
+			Bitmap check = (holder.asked ? (holder.restricted ? getFullCheckBox() : getOffCheckBox())
+					: getOnDemandCheckBox());
 
 			// Set data
 			holder.tvRestriction.setText(listLocalizedTitle.get(position));
@@ -1140,8 +1139,8 @@ public class ActivityMain extends ActivityBase implements OnItemSelectedListener
 						holder.restricted = true;
 						holder.asked = true;
 					}
-					PrivacyManager.setSetting(0, templateName, (holder.restricted ? "true" : "false")
-							+ "+" + (holder.asked ? "asked" : "ask"));
+					PrivacyManager.setSetting(0, templateName, (holder.restricted ? "true" : "false") + "+"
+							+ (holder.asked ? "asked" : "ask"));
 
 					Bitmap check = holder.asked ? (holder.restricted ? getFullCheckBox() : getOffCheckBox())
 							: getOnDemandCheckBox();