+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Changelog
 * Fixed select/unselect all ([issue](/../../issues/1390))
 * Fixed some more repeated asking on demand, thanks @[jpeg729](https://github.com/jpeg729)
 * Improved tutorial/info background ([issue](/../../issues/1391))
+* Added template for on demand restricting, thanks @[jpeg729](https://github.com/jpeg729) ([issue](/../../issues/1376))
 * Updated German translation
 * Updated French translation
 * Updated Lithuanian translation